refactor: standardize timezone parameter type to float across input validation and calendar conversion functions

+++ tuvi_mcp/_calendar.py
@@ -11,7 +11,7 @@ unified astronomical engine (Ho Ngọc Đức / Meeus, UTC+7).
 from __future__ import annotations
 
 
-def convert_solar_to_lunar(day: int, month: int, year: int, timezone: int = 7) -> dict:
+def convert_solar_to_lunar(day: int, month: int, year: int, timezone: float = 7.0) -> dict:
     """Convert a Solar date (Dương lịch) to the corresponding Lunar date (Âm lịch)."""
     from ._lunar_calendar.util.VnCalendarUtil import solar_to_lunar_vn
 
@@ -28,7 +28,7 @@ def convert_solar_to_lunar(day: int, month: int, year: int, timezone: int = 7) -
         return {"error": f"Failed to convert Solar to Lunar: {str(e)}"}
 
 
-def convert_lunar_to_solar(day: int, month: int, year: int, is_leap: bool = False, timezone: int = 7) -> dict:
+def convert_lunar_to_solar(day: int, month: int, year: int, is_leap: bool = False, timezone: float = 7.0) -> dict:
     """Convert a Lunar date (Âm lịch) to the corresponding Solar date (Dương lịch).
 
     Leap-month validation uses the engine's own helpers before delegating
